WIP: use 12 bits of precision for clipping textures

+++ src/GPU3D.h
@@ -33,7 +33,7 @@ struct Vertex
 {
     s32 Position[4];
     s32 Color[3];
-    s16 TexCoords[2];
+    s32 TexCoords[2];
 
     bool Clipped;
